Make DAG paused by default
As the docs team was going through the onboarding workflow, we noticed this detail. This seems like it would be jarring both to:

- Trial users, who suddenly are spending credits without even going to see their DAG run.
- Experienced Airflow users, who tend to manually unpause DAGs that are still being tested.

I think removing this and requiring onboarding users to trigger the DAG themselves, even though it's one more step, would be more user-friendly and consistent with the workflows they expect to complete in Astro.